cleanup some config

+++ pokebedrock/form/moderate.go
@@ -113,7 +113,7 @@ func (c CreateInfliction) Submit(sub form.Submitter, _ *world.Tx) {
 			return
 		}
 
-		expiry = time.Now().UnixMilli() + int64(time.Minute.Milliseconds())*int64(exp)
+		expiry = time.Now().UnixMilli() + time.Minute.Milliseconds()*int64(exp)
 	}
 
 	reason := c.Reason.Value()
